+++ hw/components/tty/tty_device.cpp
@@ -131,7 +131,7 @@ tty_device::~tty_device ()
     close_ttys ();
 }
 
-void tty_device::write (unsigned int ofs, unsigned char be, unsigned char *data, bool &bErr)
+void tty_device::write (unsigned long ofs, unsigned char be, unsigned char *data, bool &bErr)
 {
     unsigned char           val, tty, be_ok = true;
 
@@ -160,18 +160,18 @@ void tty_device::write (unsigned int ofs, unsigned char be, unsigned char *data,
     {
         printf ("Bad %s::%s ofs=0x%X, be=0x%X, data=0x%X-%X!\n",
                 name (), __FUNCTION__, (unsigned int) ofs, (unsigned int) be,
-                (unsigned int) *((unsigned int *)data + 0), (unsigned int) *((unsigned int *)data + 1));
+                (unsigned int) *((unsigned long*)data + 0), (unsigned int) *((unsigned long*)data + 1));
         exit (1);
     }
     bErr = false;
 }
 
-void tty_device::read (unsigned int ofs, unsigned char be, unsigned char *data, bool &bErr)
+void tty_device::read (unsigned long ofs, unsigned char be, unsigned char *data, bool &bErr)
 {
     int             i;
 
-    *((unsigned int *)data + 0) = 0;
-    *((unsigned int *)data + 1) = 0;
+    *((unsigned long *)data + 0) = 0;
+    *((unsigned long *)data + 1) = 0;
 
     switch (ofs)
     {
